Safari-609.1.11.1/LayoutTests/http/tests/media/modern-media-controls/status-support/status-support-loading-expected.txt

Testing the TimeControl behavior for the loading state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


TimeControl should not be in loading state by default.
PASS mediaController.controls.timeControl.loading is false

Media is stalled, TimeControl should now be in the loading state.
PASS mediaController.controls.timeControl.loading is true

PASS successfullyParsed is true

TEST COMPLETE
